EZP-27458: Restructured Aggregation API integration tests
