+++ eZ/Publish/API/Repository/Tests/SearchService/Aggregation/LanguageTermAggregationTest.php
@@ -1,0 +1,33 @@
+<?php
+
+/**
+ * @copyright Copyright (C) eZ Systems AS. All rights reserved.
+ * @license For full copyright and license information view LICENSE file distributed with this source code.
+ */
+declare(strict_types=1);
+
+namespace eZ\Publish\API\Repository\Tests\SearchService\Aggregation;
+
+use eZ\Publish\API\Repository\Tests\SearchService\Aggregation\DataSetBuilder\TermAggregationDataSetBuilder;
+use eZ\Publish\API\Repository\Values\Content\Query\Aggregation\LanguageTermAggregation;
+
+final class LanguageTermAggregationTest extends AbstractAggregationTest
+{
+    public function dataProviderForTestFindContentWithAggregation(): iterable
+    {
+        $aggregation = new LanguageTermAggregation('language');
+
+        $builder = new TermAggregationDataSetBuilder($aggregation);
+        $builder->setExpectedEntries([
+            'eng-US' => 16,
+            'eng-GB' => 2,
+        ]);
+
+        $builder->setEntryMapper([
+            $this->getRepository()->getContentLanguageService(),
+            'loadLanguage',
+        ]);
+
+        yield $builder->build();
+    }
+}
